requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
sqlalchemy==2.0.22
passlib==1.7.4
python-jose==3.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-asyncio==0.21.0
requests==2.32.0